requirements.txt
# sofifinmodeltool — dependencies
# Install: pip install -r requirements.txt

camelot-py[cv]==0.11.0
pdfplumber==0.11.4
openpyxl==3.1.5
openai>=1.40.0
streamlit>=1.38.0
pandas>=2.1.0
thefuzz[speedup]==0.22.1
tabulate>=0.9.0          # for DataFrame.to_markdown
python-dotenv>=1.0.0     # .env file loading
fpdf2>=2.7.0             # test PDF generation (smoke test only)
